LayoutTests/platform/glib/imported/w3c/web-platform-tests/navigation-api/navigation-history-entry/entries-after-navigations-in-multiple-windows-expected.txt



FAIL navigation.entries() behavior when multiple windows navigate. assert_equals: expected 2 but got 1
